Added few more questions.

+++ dataStructures/basicsAndMath/ChampagneTower.java
@@ -1,6 +1,6 @@
 /**
  * Created on:  Oct 28, 2020
- * Questions: https://leetcode.com/explore/challenge/card/october-leetcoding-challenge/562/week-4-october-22nd-october-28th/3508/
+ * Questions: https://leetcode.com/problems/champagne-tower/
  */
 
 public class ChampagneTower {
@@ -11,7 +11,7 @@ public class ChampagneTower {
 
     public static double champagneTower(int poured, int query_row, int query_glass) {
         double[][] A = new double[102][102];
-        A[0][0] = (double) poured;
+        A[0][0] = poured;
         for (int r = 0; r <= query_row; ++r) {
             for (int c = 0; c <= r; ++c) {
                 double q = (A[r][c] - 1.0) / 2.0;
